+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/internal/JavaViolationDecorator.java
@@ -23,6 +23,7 @@ import net.sourceforge.pmd.lang.java.ast.ASTLocalVariableDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodOrConstructorDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclarator;
 import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
+import net.sourceforge.pmd.lang.java.ast.AccessNode;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
 import net.sourceforge.pmd.reporting.ViolationDecorator;
 import net.sourceforge.pmd.util.IteratorUtil;
@@ -50,7 +51,7 @@ final class JavaViolationDecorator implements ViolationDecorator {
             enclosing = javaNode.getEnclosingType();
         }
         if (enclosing == null) {
-            enclosing = javaNode.getRoot().getTypeDeclarations().first(ASTAnyTypeDeclaration::isPublic);
+            enclosing = javaNode.getRoot().getTypeDeclarations().first(it -> it.hasVisibility(AccessNode.Visibility.V_PUBLIC));
         }
         if (enclosing == null) {
             enclosing = javaNode.getRoot().getTypeDeclarations().first();