+++ docs/pages/release_notes.md
@@ -155,6 +155,13 @@ The following previously deprecated classes have been removed:
       If the current version is needed, then `Node.getTextDocument().getLanguageVersion()` can be used. This
       is the version that has been selected via CLI `--use-version` parameter.
 
+**Deprecated classes and methods**
+
+* pmd-java
+  * {% jdoc !!java::lang.java.ast.ASTClassOrInterfaceDeclaration#isPackagePrivate() %} has been deprecated.
+    Use {% jdoc java::lang.java.ast.AccessNode#hasVisibility(java::lang.java.ast.AccessNode.Visibility) %} instead,
+    which can correctly differentiate between local and package private classes.
+
 #### External Contributions
 * [#4640](https://github.com/pmd/pmd/pull/4640): \[cli] Launch script fails if run via "bash pmd" - [Shai Bennathan](https://github.com/shai-bennathan) (@shai-bennathan)
 * [#4673](https://github.com/pmd/pmd/pull/4673): \[javascript] CPD: Added support for decorator notation - [Wener](https://github.com/wener-tiobe) (@wener-tiobe)